+++ botx/middlewares/ns.py
@@ -4,15 +4,29 @@ from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
 from uuid import UUID
 
 from loguru import logger
+from pydantic import BaseConfig, BaseModel
 
 from botx import bots
-from botx.collecting import Handler
+from botx.collecting import Collector, Handler
 from botx.concurrency import callable_to_coroutine
+from botx.dependencies import models as deps
 from botx.exceptions import NoMatchFound
 from botx.middlewares.base import BaseMiddleware
 from botx.models import messages
 from botx.typing import Executor
-from botx.utils import get_name_from_callable
+from botx.utils import get_name_from_callable, optional_sequence_to_list
+
+
+class NextStepHandlerState(BaseModel):
+    """Information about next step handler."""
+
+    name: str
+    """name of handler that should be called."""
+    arguments: Dict[str, Any]
+    """arguments that should be set on message for handler."""
+
+    class Config(BaseConfig):  # noqa: D106
+        arbitrary_types_allowed = True
 
 
 class NextStepMiddleware(BaseMiddleware):
@@ -30,7 +44,9 @@ class NextStepMiddleware(BaseMiddleware):
         executor: Executor,
         bot: bots.Bot,
         functions: Union[Dict[str, Callable], Sequence[Callable]],
-        break_handler: Optional[Union[Handler, str]] = None,
+        break_handler: Optional[Union[Handler, str, Callable]] = None,
+        dependencies: Optional[Sequence[deps.Depends]] = None,
+        dependency_overrides_provider: Any = None,
     ) -> None:
         """Init middleware with required params.
 
@@ -42,19 +58,36 @@ class NextStepMiddleware(BaseMiddleware):
                 their names.
             break_handler: handler instance or name of handler that will break next step
                 handlers chain.
+            dependency_overrides_provider: object that will override dependencies for
+                this handler.
         """
         super().__init__(executor)
-        self.break_handler: Optional[Handler] = None
-        """Handler that will be used if there is a break chain message."""
+
+        dependencies = optional_sequence_to_list(
+            bot.collector.dependencies
+        ) + optional_sequence_to_list(dependencies)
+        dep_override = (
+            dependency_overrides_provider or bot.collector.dependency_overrides_provider
+        )
+        bot.state.ns_collector = Collector(
+            dependencies=dependencies, dependency_overrides_provider=dep_override
+        )
+        bot.state.ns_store = {}
+
+        bot.state.ns_break_handler = None
         if break_handler:
-            self.break_handler = (
-                break_handler
-                if isinstance(break_handler, Handler)
-                else bot.handler_for(break_handler)
-            )
-        bot.state.ns_storage = {}
-        bot.state.ns_handlers = {}
-        bot.state.ns_arguments = {}
+            if isinstance(break_handler, Handler):
+                bot.state.ns_collector.handlers.append(break_handler)
+                bot.state.ns_break_handler = break_handler.name
+            elif callable(break_handler):
+                register_function_as_ns_handler(bot, break_handler)
+                bot.state.ns_break_handler = get_name_from_callable(break_handler)
+            else:
+                bot.state.ns_collector.handlers.append(
+                    bot.collector.handler_for(break_handler)
+                )
+                bot.state.ns_break_handler = break_handler
+
         if isinstance(functions, dict):
             functions_dict = functions
         else:
@@ -70,13 +103,18 @@ class NextStepMiddleware(BaseMiddleware):
             message: incoming message.
             call_next: next executor in middleware chain.
         """
-        if self.break_handler and self.break_handler.matches(message):
-            await self.drop_next_step_handlers_chain(message)
-            await self.break_handler(message)
-            return
+        bot = message.bot
+        if bot.state.ns_break_handler:
+            break_handler = bot.state.ns_collector.handler_for(
+                bot.state.ns_break_handler
+            )
+            if break_handler.matches(message):
+                await self.drop_next_step_handlers_chain(message)
+                await break_handler(message)
+                return
 
         try:
-            next_handler = await self.lookup_next_handler_for_message(message)
+            next_handler, state = await self.lookup_next_handler_for_message(message)
         except (NoMatchFound, IndexError, KeyError, RuntimeError):
             await callable_to_coroutine(call_next, message)
             return
@@ -86,31 +124,29 @@ class NextStepMiddleware(BaseMiddleware):
             "botx: found next step handler"
         )
 
-        arguments_lists = message.bot.state.ns_arguments[key]
-        arguments = arguments_lists.pop()
-        for argument, argument_value in arguments.items():
+        for argument, argument_value in state.arguments.items():
             setattr(message.state, argument, argument_value)
         await next_handler(message)
 
     async def lookup_next_handler_for_message(
         self, message: messages.Message
-    ) -> Handler:
+    ) -> Tuple[Handler, NextStepHandlerState]:
         """Find handler in bot storage or in handlers.
 
         Arguments:
             message: message for which next step handler should be found.
 
         Returns:
-            Found handler.
+            Found handler and state with arguments for message.
         """
-        handlers: List[str] = message.bot.state.ns_storage[
+        handlers: List[NextStepHandlerState] = message.bot.state.ns_store[
             get_chain_key_by_message(message)
         ]
-        handler_name = handlers.pop()
-        try:
-            return message.bot.state.ns_handlers[handler_name]
-        except KeyError:
-            return message.bot.handler_for(handler_name)
+        handler_state = handlers.pop()
+        return (
+            message.bot.state.ns_collector.handler_for(handler_state.name),
+            handler_state,
+        )
 
     async def drop_next_step_handlers_chain(self, message: messages.Message) -> None:
         """Drop registered chain for message.
@@ -118,7 +154,7 @@ class NextStepMiddleware(BaseMiddleware):
         Arguments:
             message: message for which chain should be dropped.
         """
-        message.bot.state.ns_storage.pop(get_chain_key_by_message(message))
+        message.bot.state.ns_store.pop(get_chain_key_by_message(message))
 
 
 def get_chain_key_by_message(message: messages.Message) -> Tuple[str, UUID, UUID, UUID]:
@@ -152,16 +188,18 @@ def register_function_as_ns_handler(
         name: name for new function. Will be generated from `func` if not passed.
     """
     name = name or get_name_from_callable(func)
-    handlers = bot.state.ns_handlers
-    if name in handlers:
-        raise ValueError(f"bot ns functions already include function with {name}")
-    handlers[name] = Handler(
-        body="",
-        handler=func,
-        include_in_status=False,
-        dependencies=bot.collector.dependencies,
-        dependency_overrides_provider=bot.dependency_overrides,
-    )
+    collector: Collector = bot.state.ns_collector
+    try:
+        collector.add_handler(
+            body=name,
+            name=name,
+            handler=func,
+            include_in_status=False,
+            dependencies=collector.dependencies,
+            dependency_overrides_provider=collector.dependency_overrides_provider,
+        )
+    except AssertionError as exc:
+        raise ValueError(exc.args)
 
 
 def register_next_step_handler(
@@ -188,21 +226,17 @@ def register_next_step_handler(
         )
 
     bot = message.bot
-    handlers = bot.state.ns_handlers
+    collector: Collector = bot.state.ns_collector
     name = get_name_from_callable(func) if callable(func) else func
 
     try:
-        bot.handler_for(name)
+        collector.handler_for(name)
     except NoMatchFound:
-        if name not in handlers:
-            raise ValueError(
-                f"bot does not have registered function or handler with name {name}"
-            )
+        raise ValueError(
+            f"bot does not have registered next step handler with name {name}"
+        )
 
     key = get_chain_key_by_message(message)
 
-    store: List[str] = bot.state.ns_storage.setdefault(key, [])
-    args: List[dict] = bot.state.ns_arguments.setdefault(key, [])
-
-    store.append(name)
-    args.append(ns_arguments)
+    store: List[NextStepHandlerState] = bot.state.ns_store.setdefault(key, [])
+    store.append(NextStepHandlerState(name=name, arguments=ns_arguments))